chore: rename squid receiver contract to be generic, delete unused widget config
