+++ libs/smart-contracts/src/contracts/index.ts
@@ -4,4 +4,4 @@ export * from './collateral-bridge';
 export * from './staking-bridge';
 export * from './token';
 export * from './multisig-control';
-export * from './arbitrum-squid-receiver';
+export * from './squid-receiver';
