+++ libs/smart-contracts/src/contracts/squid-receiver.ts
